refactor(ors_control_app): move /api/backload/seed into routes/fleet.ts
Pre-install cleanup. Move the last inline route handler out of
server/index.ts and into the existing routes/fleet.ts Router factory,
where it sits next to /api/datasets/activate and /api/fleet-config* —
all three endpoints mutate the same FLEET_INTELLIGENCE config tables.

routes/fleet.ts now imports ensureBackloadAndAssetVelocityObjects from
../lib/init.js (added in the previous commit).

server/index.ts: 225 -> 128 lines. The file is now cleanly the bootstrap
+ middleware + router mount + verifySessionUtc + app.listen scaffold,
matching the holistic refactor's Phase 1 endpoint shape.

No behaviour change. tsc passes cleanly.
